requirements.txt
ultralytics 
opencv-python 
torch 
torchvision 
matplotlib 
numpy
pillow
streamlit